+++ www/casestudies/customquestions/custom_box_whisker.php
@@ -1,46 +1,59 @@
 <?php
-    include_once '../../../config.php';
-    include_once 'includes/header.php';
 
-    use LearnositySdk\Request\Init;
-    use LearnositySdk\Utils\Uuid;
+include_once '../../config.php';
+include_once 'includes/header.php';
 
-    $session_id = Uuid::generate();
+use LearnositySdk\Request\Init;
+use LearnositySdk\Utils\Uuid;
 
-    $security = [
+$session_id = Uuid::generate();
+
+$security = [
     'user_id'      => $studentid,
     'domain'       => $_SERVER['SERVER_NAME'],
-    'consumer_key' => $consumer_key,
-    'timestamp'    => gmdate('Ymd-Hi')
-    ];
+    'consumer_key' => $consumer_key
+];
 
-    $request = '{
-        "id": "custom-mathcore",
-        "name": "Custom Mathcore",
-        "course_id": "' . $courseid . '",
-        "type": "local_practice",
-        "state": "initial",
-        "session_id": "' . $session_id . '",
-        "questions": [{
-             "response_id": "custom-mathcore-response-1",
-             "type": "custom",
-             "js": "//'. $_SERVER['HTTP_HOST'] .'/casestudies/items/customquestions/custom_mathcore.js",
-             "css": "//'. $_SERVER['HTTP_HOST'] .'/casestudies/items/customquestions/custom_mathcore.css",
-             "stimulus": "Simplify following expression: <b>\\\\(2x^2 + 3x - 5 + 5x - 4x^2 + 20\\\\)</b>",
-             "is_math": true,
-             "score": 1,
-             "specs": [{
-                         "method": "isSimplified"
-                       }, {
-                         "method": "equivSymbolic",
-                         "value": "2x^2 + 3x - 5 + 5x - 4x^2 + 20"
-             }]
-         }]
-    }';
+$request = '{
+  "id": "custom-box-whisker",
+  "name": "Custom Box Whisker",
+  "type": "local_practice",
+  "state": "initial",
+  "session_id": "' . $session_id . '",
+  "questions": [
+    {
+      "response_id": "custom-box-whisker-response-1",
+      "type": "custom",
+      "js": "//' . $_SERVER['HTTP_HOST'] . '/casestudies/items/customquestions/custom_box_whisker.js",
+      "css": "//' . $_SERVER['HTTP_HOST'] . '/casestudies/items/customquestions/custom_box_whisker.css",
+      "stimulus": "Draw a <b>box &amp; whisker</b> chart for the following: <b>6, 2, 5, 3, 6, 10, 11, 6</b>",
+      "params_line_min": 0,
+      "params_line_max": 17,
+      "params_step": 0.5,
+      "params_mark_small": 0,
+      "params_mark_big": 1,
+      "params_width": 600,
+      "params_height": 150,
+      "params_range_1": 2,
+      "params_range_2": 14,
+      "params_quartile_1": 4,
+      "params_median": 6,
+      "params_quartile_3": 10,
+      "params_box1_color": "#bbbbbb",
+      "params_box2_color": "#999999",
+      "valid_range_1": 2,
+      "valid_range_2": 11,
+      "valid_quartile_1": 4,
+      "valid_median": 6,
+      "valid_quartile_3": 8.5,
+      "score": 1
+    }
+  ]
+}';
 
 
-    $init = new Init('questions', $security, $consumer_secret, $request);
-    $signedRequest = $init->generate();
+$init = new Init('questions', $security, $consumer_secret, $request);
+$signedRequest = $init->generate();
 
 ?>
 
@@ -54,35 +67,35 @@
 
 <div class="jumbotron section">
     <div class="toolbar">
-        <ul class="list-inline">
+        <ul class="list-inline">            
             <li data-toggle="tooltip" data-original-title="Preview API Initialisation Object"><a href="#"  data-toggle="modal" data-target="#initialisation-preview"><span class="glyphicon glyphicon-search"></span></a></li>
             <li data-toggle="tooltip" data-original-title="Visit the documentation"><a href="http://docs.learnosity.com/assessment/questions/knowledgebase/customquestions" title="Documentation"><span class="glyphicon glyphicon-book"></span></a></li>
             <li data-toggle="tooltip" data-original-title="Toggle product overview box"><a href="#"><span class="glyphicon glyphicon-chevron-up jumbotron-toggle"></span></a></li>
         </ul>
-    </div>
+    </div>     
     <div class="overview">
-        <h1>Custom Question - Mathcore</h1>
-        <p>Math Custom question using Learnosity Mathcore</p>
+        <h1>Custom Question - Box Whisker</h1>
+        <p>Demostrates the implementation of a Custom question with an interactive and more complex UI.</p>
     </div>
 </div>
 
 <div class="section">        
     <div class="row">
         <div class="question-container">
-            <span class="learnosity-response question-custom-mathcore-response-1"></span>
+            <span class="learnosity-response question-custom-box-whisker-response-1"></span>
             <div class="custom-score"><strong>Score: </strong> <span id="question_score">0</span> / <span id="question_max_score">0</span></div>            
             <button class="btn btn-primary pull-right" id="validate_question">Check Answer</button>
         </div>
     </div>
 </div>    
 
-<script src="//questions.learnosity.com"></script>
+<script src="<?php echo $url_questions; ?>"></script>
 <script>
-
+   
     var questionsApp = window.questionsApp = LearnosityApp.init(<?php echo $signedRequest; ?>,  {
         errorListener: window.widgetApiErrorListener,
         readyListener: function () {
-            var question = questionsApp.question('custom-mathcore-response-1');
+            var question = questionsApp.question('custom-box-whisker-response-1');
 
             updateScores(question);
