+++ www/casestudies/customquestions/custom_mathcore.js
