+++ www/casestudies/customquestions/custom_percentage_bar.js
