+++ www/casestudies/customquestions/custom_shorttext.js
